moved inject logic to core utils
